+++ farmacia_front/src/app/app.routes.ts
@@ -1,21 +1,20 @@
 import { NgModule } from '@angular/core';
-import { RouterModule,Routes } from '@angular/router';
-import {ProveedorComponent} from './proveedor/proveedor';
+import { RouterModule, Routes } from '@angular/router';
+import { ProveedorComponent } from './proveedor/proveedor';
 import { HomeComponent } from './productos/productos';
-import { LoginComponent } from './login/login';
+import { LoginComponent } from './login/login.component';
 import { Camfarm } from './camfarm/camfarm';
 
-
 export const routes: Routes = [
-    { path: 'proveedores', component: ProveedorComponent },
-    { path: 'productos', component: HomeComponent },
-    {path: 'login', component:LoginComponent},
-    { path: 'camfarm', component: Camfarm },
-    { path: '', redirectTo: 'login', pathMatch: 'full' }
+  { path: 'proveedores', component: ProveedorComponent },
+  { path: 'productos', component: HomeComponent },
+  { path: 'login', component: LoginComponent },
+  { path: 'camfarm', component: Camfarm },
+  { path: '', redirectTo: 'login', pathMatch: 'full' },
 ];
 
 @NgModule({
   imports: [RouterModule.forRoot(routes)],
-  exports: [RouterModule]
+  exports: [RouterModule],
 })
 export class AppRoutingModule {}
